net: Move protocol list code to net_structrw.c.
This is slightly out of place because these functions aren't for
writing "structs" as such. But they still fit in better here than
in net_common.c since they aren't really just part of the common
connection code, and will be needed to include protocol list inside
query response packets.
